Move classes to client package
