+++ semester3/hw3/3.2/build.gradle
@@ -5,7 +5,7 @@ apply plugin: 'java'
 apply plugin: 'application'
 
 sourceCompatibility = 1.8
-mainClassName = 'com.rprtr258.Client'
+mainClassName = 'com.rprtr258.client.Main'
 
 repositories {
     mavenCentral()
