+++ semester3/hw3/3.2/src/main/java/com/rprtr258/client/Bullet.java
@@ -1,4 +1,4 @@
-package com.rprtr258;
+package com.rprtr258.client;
 
 import javafx.geometry.Point2D;
 import javafx.scene.canvas.GraphicsContext;
